+++ build-researcher-loot-locally.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+rm -rf Config
+mkdir Config
+go run .
+cp -r Config-Shared/* Config
+cp -r Config-Researcher-Loot/* Config