+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.1.0] - 2022-11-11
 
 - add hotbox to traders' regular inventory
+- add new researcher-loot edition
+- add researcher schematics to quest loot
 
 ## [2.0.0] - 2022-09-17